+++ qiita-control-plane/src/qiita_control_plane/routes/__init__.py
@@ -8,6 +8,7 @@ from .alignment import alignment_definition_router, alignment_router
 from .auth import router as auth_router
 from .biosample import biosample_router as biosample_top_level_router
 from .biosample import router as biosample_router
+from .exported_identifier import router as exported_identifier_router
 from .host_filter_profile import router as host_filter_profile_router
 from .prep_protocol import router as prep_protocol_router
 from .prep_sample import router as prep_sample_router
@@ -47,6 +48,7 @@ api_router.include_router(alignment_router)
 api_router.include_router(prep_protocol_router)
 api_router.include_router(prep_sample_router)
 api_router.include_router(prep_sample_study_router)
+api_router.include_router(exported_identifier_router)
 api_router.include_router(study_router)
 api_router.include_router(upload_router)
 api_router.include_router(user_router)